toxic-news-pipeline/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.1.1
pandas==2.1.3
feedparser==6.0.10
trafilatura==2.0.0
requests==2.31.0
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv
transformers==4.41.0 
sentencepiece
protobuf>=4.23.0
matplotlib

